+++ test/help/help-commands.test.js
@@ -4,9 +4,10 @@ const { run } = require('../utils/test-utils');
 const helpHeader = 'The build tool for modern web applications';
 
 describe('commands help', () => {
-    it('throws error if supplied as an argument for subcommands', () => {
-        const { stderr } = run(__dirname, ['serve', 'help'], false);
-        expect(stderr).toContain('Unknown argument: help');
+    it('shows help for subcommands', () => {
+        const { stderr, stdout } = run(__dirname, ['serve', 'help'], false);
+        expect(stderr).toBeFalsy();
+        expect(stdout).toContain('webpack s | serve');
     });
 
     it('shows help information with subcommands as an arg', () => {