+++ packages/generate-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/webpack/webpack-cli/compare/@webpack-cli/generate-plugin@1.0.1...@webpack-cli/generate-plugin@1.0.2) (2020-10-19)
+
+**Note:** Version bump only for package @webpack-cli/generate-plugin
+
 ## [1.0.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/generate-plugin@1.0.1-rc.1...@webpack-cli/generate-plugin@1.0.1) (2020-10-10)
 
 ### Bug Fixes